Handle potential packet splits while reading chunked footers
While processing a chunked response, the footer will end with a CRLF.
It is possible that the CR and LF characters will get split into two packets
so that the ByteBuffer ends with a CR character. This causes `state` to change
to `ALL_READ` which is incorrect. It should read the LF from the next packet and
then flip state to `ALL_READ`.